Stuff is working
I got git to work in my computer in cmd line and w/ eclipse. just needed
to restart it!

I added some functions into textToMorse(it now goes the other way!!!)
and imported it into keygetter. I modified keygetter so that if you
don't type anything the program prints "/" and if you use the 1 key to
enter in the morse code. press escape to exit program and it will print
out what you morsed as text!!

also of note clock.tick(NUMBER) does not delay NUMBER of miliseconds. it
delays a certain amount of time to keep the program from running faster
than NUMBER cycles per second. so it delays 1/NUMBER seconds. it also
returns a floating point showing the number of seconds since it was last
called.